+++ tutorial/main.cpp
@@ -43,7 +43,7 @@ int main()
     // ===================
 
     // Circuit tick method 1: Manual
-    for (unsigned short i = 0; i < 10; i++)
+    for (int i = 0; i < 10; i++)
     {
         circuit.Tick();
         circuit.Reset();
